RR Controllers good to go

+++ Project/Features/Ratings and Reviews/RatingAndReview.cs
@@ -29,7 +29,7 @@ namespace Features.Ratings_and_Reviews
             this.DispName = dispName;
         }
 
-        public RatingAndReview(string dispName, string selectedTitle, float userRating)
+        public RatingAndReview(string dispName, string selectedTitle, double userRating)
         {
             this.Rating = userRating;
             this.Review = "";
@@ -45,7 +45,7 @@ namespace Features.Ratings_and_Reviews
             this.DispName = dispName;
         }
 
-        public RatingAndReview(string dispName, string selectedTitle, float userRating, string userReview)
+        public RatingAndReview(string dispName, string selectedTitle, double userRating, string userReview)
         {
             this.Rating = userRating;
             this.Review = userReview;
